refactor freelancer profile and adding some logic
