+++ src/components/pages/FreelancerProfile/components/index.js
@@ -1,0 +1,8 @@
+export * from "./PersonalInfo";
+export * from "./FinishedJobs";
+export * from "./ResumeSection";
+export * from "./Posts";
+export * from "./SectionCard";
+export * from "./Certificates";
+export * from "./EditModal";
+export * from "./Header";
